backend/requirements.txt
africastalking==1.2.9
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coverage==7.7.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-allauth==65.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
exceptiongroup==1.2.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
schema==0.7.7
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.0
urllib3==2.3.0
